Seperate test for focus functions.
