+++ cameramodule/src/test/java/com/benine/ipcameracontrol/IpcameraTest.java
@@ -126,140 +126,7 @@ public class IpcameraTest {
     mockServerClient.verify(request, VerificationTimes.once());
   }
   
-  @Test
-  public final void testGetFocusPosition() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#GF"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("gfA42"));
-
-    int res = camera.getFocus().getFocusPos();
-    
-    mockServerClient.verify(request, VerificationTimes.once());
-    
-    assertEquals(res, 2626, 0.000001);
-  }
   
-  @Test(expected = IpcameraConnectionException.class)
-  public final void testGetFocusPositionException() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#GF"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("ggA42"));
-
-    camera.getFocus().getFocusPos();
-  }
-  
-  @Test
-  public final void testSetFocus() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#AXFFFF"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("axfFFF"));
-
-    camera.getFocus().setFocusPos(2882);
-    
-    mockServerClient.verify(request, VerificationTimes.once());
-  }
-  
-  @Test
-  public final void testMoveFocus() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#F80"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("fs80"));
-
-    camera.getFocus().moveFocus(80);
-    
-    mockServerClient.verify(request, VerificationTimes.once());
-  }
-  
-  @Test
-  public final void testIsAutoFocusOn() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#D1"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("d11"));
-
-    boolean res = camera.getFocus().isAutoFocusOn();
-    
-    mockServerClient.verify(request, VerificationTimes.once());
-    assertTrue(res);
-  }
-  
-  @Test
-  public final void testIsAutoFocusOff() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#D1"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("d10"));
-
-    boolean res = camera.getFocus().isAutoFocusOn();
-    
-    mockServerClient.verify(request, VerificationTimes.once());
-    assertFalse(res);
-  }
-  
-  @Test(expected = IpcameraConnectionException.class)
-  public final void testIsAutoFocusOffException() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#D1"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("K10"));
-
-    camera.getFocus().isAutoFocusOn();
-  }
-  
-  @Test
-  public final void testSetAutoFocusOff() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#D10"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("d10"));
-
-    camera.getFocus().setAutoFocusOn(false);
-    
-    mockServerClient.verify(request, VerificationTimes.once());
-  }
-  
-  @Test
-  public final void testSetAutoFocusOn() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#D11"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("d11"));
-
-    camera.getFocus().setAutoFocusOn(true);
-    
-    mockServerClient.verify(request, VerificationTimes.once());
-  }
   
   @Test
   public final void testSetAutoIrisOff() throws CameraConnectionException {
